GMAT path parsing (#23)
* refactor: move import_lib.py contents into __init__.py, remove extra newline

* chore: auto-update VCS directory mappings
